Grant request (#295)
* Create README.md

* Add files via upload

* Add files via upload

* Update and rename projects/ST7735S-Breakout/PCB_PCB_ST7735S Breakout_2024-01-01.json to projects/ST7735S-Breakout/src/PCB_PCB_ST7735S Breakout_2024-01-01.json

* Add files via upload

Please note this is not the most recent update-I have since removed the unintentional extra JLCJLCJLCJLC in the top layer. Everything else is as is.

* Added uploaded gerber.

Please note this is not the most recent version-I have since fixed the major error (extra JLCJLCJLCJLC in top layer). Everything else is as shown.

* Added missing schematic file

* reorg

---------

Co-authored-by: Karmanyaah Malhotra <karmanyaah.gh@malhotra.cc>
Co-authored-by: Karmanyaah Malhotra <karmanyaah.git@malhotra.cc>